drm/imx: imx8mp-ldb: fix check on LDB valid clock
Remove wrong mode-fixup on the LDB driver that limited the pixel clocks
to 74.25MHz in single channel mode and 148.5MHz in dual channel mode.

According to the iMX8MP Reference manual, the LDB peripheral supports:
"
- Single channel (4 lanes) output at up to 80MHz pixel clock and LVDS clock.
  This supports resolutions up to 1366x768p60.
- Dual asynchronous channels (8 data, 2 clocks). This is intended for a single
  panel with two interfaces, transferring across two channels (even pixel/odd
  pixel). This is supported at up to 160MHz pixel clock, which is up to 80MHz
  LVDS clock (due to 2 pixels per LVDS clock). This supports resolutions
  above 1366x768p60, up to 1080p60.
"

We contacted NXP trying to find the real reason for this seemingly
random limitation, but we didn't get any response.

The forced pixel clock values were removed and now it's possible to
drive LVDS displays that have pixel clock requirements out of the previously
limited range, but that are still within the hardware limitations.
A check was added to ensure that clocks out of range are denied.

With this change, it is now possible to use this display:
- LVDS display (CMT1210A07R00 800x600 - 40MHz pixel clock)

Also, the Toradex 10" Capacitive touch display (68.93MHz pixel clock)
continues to work.

Upstream-Status: Pending

The file where this change was applied doesn't exist upstream.

Signed-off-by: Rafael Beims <rafael.beims@toradex.com>