clk: imx: pll14xx: consolidate rate calculation
The PLL driver has support for two different PLLs: The pll1416x and
the pll1443x. The latter has support for an additional kdiv value.
recalc_rate can be the same calculation when kdiv is assumed to be zero
for the PLL which doesn't support that value.

Upstream-Status: Backport [53990cf9d5b489fc0ec08e5c4df7139fc311a824]

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
Reviewed-by: Abel Vesa <abel.vesa@nxp.com>
Link: https://lore.kernel.org/r/20220304125256.2125023-5-s.hauer@pengutronix.de
Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
[Rafael: fixed conflicts in the clk_pll1443x_recalc_rate function caused
by some additions on the downstream branch in commit 9425151a59a4. The
changes in the downstream branch implement the same functionality of
this commit but without the other improvements in the patch series.
Additionally, the change in the downstream branch adds a function to
change the "k" value of the PLL which gets exposed through sysfs. This
function was also fixed to be compatible with the changes in the patch
series.]
Signed-off-by: Rafael Beims <rafael.beims@toradex.com>