arm64: dts: colibri-imx8x: Add PMIC thermal zone
Add PMIC thermal zone as is was present on both Colibri-iMX8x/BSP5 and
in MEK board.

Without this the thermal sensor driver fails to probe on colibri-imx8x
with i.MX8X DX SOC, this is happening because imx8dx-colibri.dtsi
creates an incomplete thermal zone node for the PMIC.

Upstream-Status: Pending

This should be send upstream, however the first batch of DTS files is
under review, therefore it seems better to just postpone it.

Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>