ata: ahci-imx: Do not disable APB clock
There is a clock dependency between the ahci-imx driver and the PCIe
driver on i.MX8 which causes the PCIe link to never come up. It happens
every 20-100 reboots when the SATA driver probes before PCIe. By never
disabling the APB clock in the SATA driver this issue goes away. In BSP5
we compiled the SATA driver as a module and the SATA driver was probing
after the PCIe driver, therefore we could not see the issue there.

Upstream-Status: Inappropriate [other]

This is an ugly workaround that should not land upstream. We try to get
help from NXP regarding this issue to find a proper solution.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>